Fix app dir vanilla-extract support (#48306)
Closes #48222.

This PR introduces a new utility function that checks the resource path,
module type, and loaders to replace the existing CSS regex in the Flight
loaders and plugins. The new function is more robust and can be extended
to correctly support other libraries such as `vanilla-extract`.

cc @SuttonJack and @SukkaW (this could be related to the other issue
we've discussed too).